avslutt behandling knapp bruker secondary variant & alignes med send til beslutter
