+++ src/components/behandling/førstegangsbehandling/7-send-og-godkjenn/FørstegangsVedtakKnapper.module.css
@@ -1,0 +1,8 @@
+.knapper {
+    margin-bottom: var(--a-spacing-8);
+}
+
+.godkjentWrapper {
+    display: flex;
+    justify-content: end;
+}
