--- src/components/behandling/førstegangsbehandling/7-send-og-godkjenn/FørstegangsVedtakSend.module.css
@@ -1,4 +1,0 @@
-.knapper {
-    justify-self: end;
-    margin-top: 2rem;
-}
