+++ src/components/behandling/førstegangsbehandling/FørstegangsVedtak.tsx
@@ -3,10 +3,9 @@ import { FørstegangsVedtakBegrunnelse } from './1-begrunnelse/FørstegangsVedta
 import { FørstegangsVedtakResultat } from './2-resultat/FørstegangsVedtakResultat';
 import { FørstegangsVedtakBrev } from './6-brev/FørstegangsVedtakBrev';
 import { Separator } from '../../separator/Separator';
-import { FørstegangsVedtakSend } from './7-send-og-godkjenn/FørstegangsVedtakSend';
+import { FørstegangsVedtakKnapper } from './7-send-og-godkjenn/FørstegangsVedtakKnapper';
 import { FørstegangsVedtakBarnetillegg } from './5-barn/FørstegangsVedtakBarnetillegg';
 import { FørstegangsVedtakProvider } from './context/FørstegangsVedtakContext';
-import FørstegangsvedtakAvbrytBehandling from './8-avbryt-behandling/FørstegangsvedtakAvbrytBehandling';
 import { FørstegangsVedtakTiltak } from './4-tiltak/FørstegangsVedtakTiltak';
 
 import style from './FørstegangsVedtak.module.css';
@@ -27,8 +26,7 @@ export const FørstegangsVedtak = () => {
             <FørstegangsvedtakAvslagsgrunner />
             <FørstegangsVedtakBarnetillegg />
             <FørstegangsVedtakBrev />
-            <FørstegangsVedtakSend />
-            <FørstegangsvedtakAvbrytBehandling />
+            <FørstegangsVedtakKnapper />
         </FørstegangsVedtakProvider>
     );
 };
