+++ src/components/saksoversikt/avsluttBehandling/AvsluttBehandling.tsx
@@ -61,7 +61,7 @@ const AvsluttBehandling = (props: {
                 />
             )}
             <Button
-                variant="danger"
+                variant="secondary"
                 type="button"
                 size={props.button?.size ?? 'small'}
                 onClick={() => setVilAvslutteBehandling(true)}
